+++ docs/source/News.rst
@@ -5,6 +5,13 @@
 This document describes the releases of :mod:`repoze.what.plugins.sql`.
 
 
+Unreleased
+==========
+
+Made compatible with SQLAlchemy 0.7b2, by replacing import from
+:mod:`sqlalchemy.exceptions` with :mod:`sqlalchemy.exc`.
+
+
 Version 1.0 Final (2010-03-11)
 ==============================